LLAMA-2672 : Issues waking from NSM
Reason for change: LLAMA-2672 is specifically about control manager
trying to message when Power Manager when Power Manager isn't ready.
We are changing the approach to handling NSM wakeups, to clear a
lot of patches that have been put in place around handling a spurious
interrupt.

Test Procedure: Run full regression testing on NSM test cases.
Including the tests for LLAMA-2104 around muting and un-muting the
mics while the device is in NSM

Risks: Must test all NSM to be sure these changes are good

Signed-off-by: Jason Thomson <jason_thomson@comcast.com>
Change-Id: I3cde96b3b578eeed7103bbbc92305ff1ac12c6d6